Fix entity blocking for tree deeper than one node
This commit removes the strange behavior where immediate children of grid view orderings had an special treatment. Their blocking was deactivated with no apparent reason. This happened to work for simple trees, but failed on more complicated trees. The more general approach is to set the blocking of the children GFS to the local orderings.