Set proper blocking of interleaved nodes by entity
The topological associativity forest is technically an interleaving of
lower nodes so that the index of the entity raises to the root node of the trees.
In this sense, the blocking assigned to the leaf nodes should raise together with the interleaving process.
What this means is: the blocking assigned to the leaf nodes must be all the same, and it will determine if the resulting index for indices is blocked or merged.